Everything implemented so far passing all tests

+++ src/constants.rs
@@ -10,7 +10,7 @@ pub const EPLITE_MAGIC_HEADER_V1: &[u8; 16] = b"EPLite format 1\0";
 pub const SQLITE_MAGIC_HEADER_V3: &[u8; 16] = b"SQLite format 3\0";
 
 /// Valid page size range
-pub const PAGE_SIZE_RANGE: RangeInclusive<u16> = 512..=65536;
+pub const PAGE_SIZE_RANGE: RangeInclusive<u32> = 512..=65536;
 
 /// Default page size (4096 bytes)
 pub const DEFAULT_PAGE_SIZE: u32 = 4096;
